requirements.txt
python-dotenv
supabase
feedparser
beautifulsoup4
requests
python-dateutil